[539021] Prevent LockStatePrefetcher from getting CDOObject if not needed
Revision manager is associated to the CDOSession. The
LockStatePrefetcher is notified for every revisions for all CDOView of
the CDOSession
Consequently, for view with other time stamp that the one for which the
revision is loaded, the CDOObject may not exist or the revision may not
be loaded. In the former, a CDOObjectNotFoundException could be thrown
and in the latter the revision could be loaded whereas it is not needed.
To avoid that, we check that the revision, for which the LockState would
be loaded, is effectively part of the revisions of this CDOView. If not
we do nothing.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=539021
Change-Id: Ib80290e67d25ee662d3bbebe2b7e35b04ccdb4f9
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>